==> 1. Ophalen van de Manifesto Project-database.ipynb <==
## 1. Ophalen van de Manifesto Project-database

Dit notebook haalt de eerst algemene, geaggregeerde dataset van het Manifesto Project op, en daarna de coderingen van vijftien Westerse landen op zinsniveau. Voor dat laatste is een API key vereist. Die verkrijg je door een [account aan te maken](https://manifesto-project.wzb.eu/signup) op de website van het Manifesto Project.

import wget

wget.download('https://manifesto-project.wzb.eu/down/data/2023a/datasets/MPDataset_MPDS2023a.csv', 'data')

API_KEY = ''

import datetime

import requests
import pandas as pd
from tqdm.auto import tqdm

data = pd.read_csv('data/MPDataset_MPDS2023a.csv', low_memory=False)
data.head(5)

SELECTED_COUNTRIES = ['United States', 'Canada', 'Spain', 'Belgium', 'Netherlands', 'France', 'Denmark',
 'Italy', 'Austria', 'Germany', 'Switzerland', 'Sweden', 'United Kingdom', 'Portugal', 'Finland']

data = data[data.countryname.isin(SELECTED_COUNTRIES)]
data['date'] = pd.to_datetime(data.edate, dayfirst=True)
data = data[data.date > datetime.datetime(2000,1,1)]

def manifesto_sentences(manifesto_id):
    url = "https://manifesto-project.wzb.eu/api/v1/texts_and_annotations?api_key={}&keys[]={}&version=2023-1"
    url = url.format(API_KEY, manifesto_id)
    response = requests.get(url).json()
    return response

rows = []
for row in tqdm(data.itertuples(), total=len(data)):
    manifesto_id = str(row.party) + '_' + str(row.date)[:7].replace('-', '')
    response = manifesto_sentences(manifesto_id)
    if 'items' in response:
        for item in response['items']:
            for i in item['items']:
                i['party'] = row.partyabbrev
                i['date'] = row.date
                i['country'] = row.countryname
                rows.append(i)

manifestos = pd.DataFrame(rows)

selected_cols = ['101', '102', '103', '104', '105', '106', '107', '108', '109',
                 '110', '201', '202', '203', '204', '301', '302', '303', '304',
                 '305', '401', '402', '403', '404', '405', '406', '407', '408',
                 '409', '410', '411', '412', '413', '414', '415', '416', '501', 
                 '502', '503', '504', '505', '506', '507', '601', '602', '603',
                 '604', '605', '606', '607', '608', '701', '702', '703', '704',
                 '705', '706']

manifestos.loc[manifestos['party'] == 'D’66', 'party'] = 'D66'
manifestos['code'] = manifestos.cmp_code.apply(lambda x: x[:3] if str(x[:3]) in selected_cols else '000')

manifestos[['text', 'party', 'country', 'date', 'code'
           ]].to_csv('data/manifestos_2000_2022_coded_sents.csv', index=False)
