==> 5. Analyse van onderscheidende woorden.ipynb <==
## 5. Analyse van onderscheidende woorden

Ten slotte vergelijken we in dit notebook de woordkeuzes van de rechtse partijen in hun programma’s. Welke woorden karakteriseren bijvoorbeeld de VVD als het gaat over migratie? 

Dat berekenden we aan de hand van chikwadraatoets, waarbij we de frequentie van een woord in een programma afzetten tegen van de frequentie van datzelfde woord in de andere programma’s van rechtse partijen van dat jaar. 

We nemen voor deze analyse alleen zelfstandige en bijvoegelijke naamwoorden mee, en woorden die minder dan vijf keer voorkwamen hebben we eruit gefilterd. 

Om het aantal onderwerpen behapbaar te maken, brachten we de 56 categorieën van het Manifesto Project terug tot 12 onderwerpen. Daarvoor gebruikten we de groepering van onderzoekers van de VU.   

import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt

from collections import Counter

from helpers.keyword_analysis import significant_words

%config InlineBackend.figure_format='retina'
%matplotlib inline
mpl.style.use('../groene.mpl')

right_wing_parties = ['CDA', 'VVD', 'PVV', 'FVD', 'BBB', 'JA21', 'NSC', 'FvD', 'BVNL']
sents = pd.concat([pd.read_csv('data/manifestos_nl_2023_coded_sents.csv'),
                   pd.read_csv('data/manifestos_2000_2022_coded_sents.csv')])

data = sents[(sents.country == 'Netherlands')&(sents.party.isin(right_wing_parties))].reset_index(drop=True)
proportions = pd.read_csv('data/manifestos_2000_2023_proportions.csv')
proportions['year'] = proportions.date.apply(lambda x: int(x[:4]))

import spacy

nlp = spacy.load('nl_core_news_lg')

from tqdm.auto import tqdm

def remove_non_printable(input_string):
    cleaned_string = "".join(c for c in input_string if c.isprintable())
    return cleaned_string

texts = list(data.text)
texts = [remove_non_printable(text) for text in tqdm(texts)]
processed_texts = [text for text in tqdm(nlp.pipe(texts, 
                                              n_process=4, # four threads
                                              disable=["ner",
                                                       "parser"]),
                                          total=len(texts))]

tokenized_texts = [[word.lemma_.lower() for word in text 
                    if word.pos_ in ['NOUN', 'ADJ'] and len(word.lemma_.lower()) > 2] 
                   for text in processed_texts]

data['tokenized_texts'] = tokenized_texts

party_dict = dict()

## Eerst: onderscheidende woorden algemeen

issue = 'Algemeen'

aflatten = lambda t: [item for sublist in t for item in sublist]

data['year'] = data.date.apply(lambda x: int(x[:4]))

filter_words = ['forum', 'democratie', 'ers']
for idx,(year, year_df) in enumerate(data[(data.year > 2003)].groupby('date')):
    parties = set(year_df.party)
    for i, party in enumerate(parties):
        if party not in party_dict:
            party_dict[party] = dict()
        if issue not in party_dict[party]:
            party_dict[party][issue] = dict()
        source = year_df[year_df.party == party].tokenized_texts.values
        target = year_df[year_df.party != party].tokenized_texts.values
        results = pd.DataFrame(significant_words(source, target))
        results = results[(~results.word.str.contains('-|x'))&(~results.word.isin(filter_words))]
        results = results[(results.freq_c1>=5)&(results.freq_c2>0)].sort_values('chi', ascending=False)
        results = results[results.chi_p < 0.05]
        if len(results) > 0:
            party_dict[party][issue][year[:4]] = list(results.word.values[:5])

party_dict['VVD']

party_dict['BBB']

party_dict['NSC']

## Dan: onderscheidende woorden per onderwerp

code2issue = pd.read_csv('https://raw.githubusercontent.com/vanatteveldt/2023-manifestos-nl/main/data/raw/cmp_topics.csv')
code2issue = dict(zip(code2issue['cmp'], code2issue['label']))

data['issue'] = data.code.apply(lambda x: code2issue[x] if x in code2issue else '')

issue2include = ['Migratie', 'Bestaanszekerheid', 'Veiligheid', 'Milieu', 'Economie', 'Normen']

for issue in issue2include:
    aflatten = lambda t: [item for sublist in t for item in sublist]

    data['year'] = data.date.apply(lambda x: int(x[:4]))

    for idx,(year, year_df) in enumerate(data[(data.year > 2003)&(data.issue == issue)].groupby('date')):
        parties = set(year_df.party)
        for i, party in enumerate(parties):
            if party not in party_dict:
                party_dict[party] = dict()
            if issue not in party_dict[party]:
                party_dict[party][issue] = dict()

            source = year_df[year_df.party == party].tokenized_texts.values
            target = year_df[year_df.party != party].tokenized_texts.values
            results = pd.DataFrame(significant_words(source, target, tests=('llr', 'chi')))
            results = results[(~results.word.str.contains('-|x'))&(~results.word.isin(filter_words))]
            results = results[(results.freq_c1>=5)&(results.freq_c2>0)].sort_values('chi', ascending=False)
            results = results[results.chi_p < 0.05]
            if len(results) > 0:
                party_dict[party][issue][year[:4]] = list(results.word.values[:5])

party_dict['VVD']['Migratie']

party_dict['BBB']['Migratie']

party_dict['FvD']['Migratie']

party_dict['NSC']['Migratie']

party_dict['VVD']['Milieu']
